% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadConfigs.R
\name{loadConfigBioinfo}
\alias{loadConfigBioinfo}
\title{Load config for the bioinfo server (or any other server than rbalhpc05)}
\usage{
loadConfigBioinfo()
}
\description{
Load config for the bioinfo server (or any other server than rbalhpc05)
}

