% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadConfigs.R
\name{loadConfigRbalhpc05}
\alias{loadConfigRbalhpc05}
\title{Load config for the rbalhpc05 server (which runs Torque/PBS)}
\usage{
loadConfigRbalhpc05()
}
\description{
Load config for the rbalhpc05 server (which runs Torque/PBS)
}

