% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeClusterFunctionsRemoteTorque.R
\name{makeClusterFunctionsRemoteTorque}
\alias{makeClusterFunctionsRemoteTorque}
\title{makeClusterFunctions for remote torque}
\usage{
makeClusterFunctionsRemoteTorque(template.file,
  list.jobs.cmd = c(QSELECT_COMMAND, "-u $USER", "-s EHQRTW"))
}
\arguments{
\item{template.file}{the same as in makeClusterFunctionTorque}

\item{list.jobs.cmd}{the same as in makeClusterFunctionTorque

The function extends makeClusterFunctions in BatchJobs. It allows submitting jobs to a remote server where the Torque/PBS system is installed, via ssh.}
}
\description{
makeClusterFunctions for remote torque
}

