% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribiosCfBrewTemplate.R
\name{ribiosCfBrewTemplate}
\alias{ribiosCfBrewTemplate}
\title{cfBrewTemplate customed for ribios}
\usage{
ribiosCfBrewTemplate(conf, template, rscript, extension)
}
\arguments{
\item{conf}{the same as in cfBrewTemplate}

\item{template}{the same as in cfBrewTemplate}

\item{rscript}{the same as in cfBrewTemplate}

\item{extension}{the same as in cfBrewTemplate

The function is adapted from cfBrewTemplate. Instead of writing to /tmp/, which can be only accessed by the local machine but not the remote machine, the function uses ribiosTempfile implemented in ribiosUtils to allow cluster-visible temporary files.}
}
\description{
cfBrewTemplate customed for ribios
}

